tests/wpt/tests/css/css-box/parsing/margin-trim.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS margin-trim: property parsing</title>
<link rel="help" href="https://www.w3.org/TR/css-box-4/#margin-trim">
<link rel="author" title="Sammy Gill" href="mailto:sammy.gill@apple.com">
<meta name="assert" content="Test parsing for the margin-trim property">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
// Individual values should get set
test_valid_value("margin-trim", "none");
test_valid_value("margin-trim", "block");
test_valid_value("margin-trim", "inline");
test_valid_value("margin-trim", "block-start");
test_valid_value("margin-trim", "block-end");
test_valid_value("margin-trim", "inline-start");
test_valid_value("margin-trim", "inline-end");

// Serialize values into either block or inline
test_valid_value("margin-trim", "block-start block-end", "block");
test_valid_value("margin-trim", "inline-start inline-end", "inline");
test_valid_value("margin-trim", "block-end block-start", "block");
test_valid_value("margin-trim", "inline-end inline-start", "inline");
test_valid_value("margin-trim", "inline-start block-start");

test_valid_value("margin-trim", "inline-end block-start block-end");
test_valid_value("margin-trim", "block-start inline-start block-end inline-end");
test_valid_value("margin-trim", "inline-end block-end inline-start block-start");

test_invalid_value("margin-trim", "block inline");
test_invalid_value("margin-trim", "block block");
test_invalid_value("margin-trim", "inline inline");
test_invalid_value("margin-trim", "block inline-start inline-end");
test_invalid_value("margin-trim", "block block-start block-end");
test_invalid_value("margin-trim", "block-start block-end block");
test_invalid_value("margin-trim", "block 10px");
test_invalid_value("margin-trim", "auto");
test_invalid_value("margin-trim", "left");
</script>
</body>
</html>
